refactor: コンポーネントのディレクトリ構成を変更
