+++ frontend/src/components/user/user-bulk-delete-dialog.tsx
