+++ frontend/src/pages/dashboard.tsx
@@ -1,5 +1,5 @@
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import { AppSidebar } from "./app-sidebar";
+import { AppSidebar } from "@/components/layout/app-sidebar";
 
 export function Dashboard() {
   return (
