+++ frontend/src/services/user-service.ts
@@ -1,4 +1,4 @@
-import type { User } from "@/components/user-table";
+import type { User } from "@/types/api";
 
 const API_BASE_URL =
   import.meta.env.VITE_API_BASE_URL || "http://localhost:8000";
